+++ includes/features/class-classtypes.php
@@ -28,7 +28,7 @@ class ClassTypes {
 	 * @since  1.0.0
 	 * @var    array    $classes    Maintains the classes list.
 	 */
-	public static $classes = [ 'mail', 'plugin', 'unknown' ];
+	public static $classes = [ 'mail', 'plugin', 'theme', 'mu', 'unknown' ];
 
 	/**
 	 * The list of classes names.
@@ -46,6 +46,8 @@ class ClassTypes {
 	public static function init() {
 		self::$classe_names['mail']    = esc_html__( 'Core', 'mailarchiver' );
 		self::$classe_names['plugin']  = esc_html__( 'Plugin', 'mailarchiver' );
+		self::$classe_names['theme']   = esc_html__( 'Theme', 'mailarchiver' );
+		self::$classe_names['mu']      = esc_html__( 'MU-Plugin', 'mailarchiver' );
 		self::$classe_names['unknown'] = esc_html__( 'Unknown', 'mailarchiver' );
 	}