+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to **MailArchiver** is documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **MailArchiver** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased - will be 1.2.0]
+### Added
+- Support for plugged (custom) `wp_mail` function (mu-plugins or themes).
+
 ## [1.1.0]
 ### Added
 - Full compatibility with WP Mail SMTP.